Refactoring into different files
